requirements-minimal.txt
# Minimal requirements to get started
httpx>=0.28.1
python-dotenv>=1.0.1
fastapi>=0.110.0
uvicorn>=0.27.1
inquirerpy>=0.3.4